Add quick node support.
Add a "main" entry to package.json so the EpicEditor script can be found
via require('epiceditor'), and put Marked (the dependency) before
EpicEditor (the library) in the generated .js and .min.js sources such
that EpicEditor is the last to touch module.exports (Marked is currently
clobbering it).

A cleaner way to do this would be to build a version of EpicEditor that
doesn't contain Marked, and allow clients to include this dependency
separately. In that case, main.js could reference the trimmed-down
version of EpicEditor (minus the Marked dependency), because any client
using EpicEditor that way can just do require('marked') beforehand.